Various fixes related to mexfiles.
These were tested using ElCapitan and Matlab 2017a.

+++ mgllib/mglPrivateGetKeyEvent.c
@@ -156,7 +156,7 @@ mxArray *makeOutputStructure(double **outptrCharCode, double **outptrKeyCode, do
   // create the output structure
   mxArray *plhs;
   const char *fieldNames[] =  {"charCode","keyCode","keyboard","when" };
-  int outDims[2] = {1, 1};
+  size_t outDims[2] = {1, 1};
   plhs = mxCreateStructArray(1,outDims,4,fieldNames);
   
   // Create the output fields
